backend/services/bitcoin_service.py
"""
Bitcoin service for YubiKey Bitcoin Seed Storage
"""

import os
import yaml
from typing import Dict, Any, List, Optional

# This is a placeholder for the actual Bitcoin service implementation
# In a real implementation, this would be extracted from bitcoin_utils.py

class BitcoinService:
    """Service for handling Bitcoin operations"""
    
    def __init__(self, strength: int = 256):
        """
        Initialize the Bitcoin service
        
        Args:
            strength: Default entropy strength in bits (128, 160, 192, 224, or 256)
        """
        # Load configuration
        with open(os.path.join(os.path.dirname(__file__), '..', 'config.yaml'), 'r') as f:
            self.config = yaml.safe_load(f)
        self.default_strength = strength
    
    def generate_mnemonic(self, strength: Optional[int] = None) -> str:
        """
        Generate a BIP39 mnemonic seed phrase
        
        Args:
            strength: Entropy strength in bits (128, 160, 192, 224, or 256)
            
        Returns:
            A BIP39 mnemonic seed phrase
        """
        # Use default strength if none provided
        if strength is None:
            strength = self.default_strength
            
        # This is a simplified implementation
        # In a real implementation, this would use the mnemonic library
        
        # Generate random entropy
        entropy = os.urandom(strength // 8)
        
        # Convert to mnemonic (simplified)
        # In a real implementation, this would use the mnemonic library
        return "example mnemonic seed phrase for demonstration purposes only"
    
    def validate_mnemonic(self, mnemonic: str) -> bool:
        """
        Validate a BIP39 mnemonic seed phrase
        
        Args:
            mnemonic: The mnemonic seed phrase to validate
            
        Returns:
            True if the mnemonic is valid, False otherwise
        """
        # This is a simplified implementation
        # In a real implementation, this would use the mnemonic library
        
        # Check if the mnemonic is valid (simplified)
        # In a real implementation, this would use the mnemonic library
        return len(mnemonic.split()) in [12, 15, 18, 21, 24]
    
    def mnemonic_to_seed(self, mnemonic: str, passphrase: str = "") -> bytes:
        """
        Convert a BIP39 mnemonic to a seed
        
        Args:
            mnemonic: The mnemonic seed phrase
            passphrase: Optional passphrase for additional security
            
        Returns:
            The seed as bytes
        """
        # This is a simplified implementation
        # In a real implementation, this would use the mnemonic library
        
        # Convert mnemonic to seed (simplified)
        # In a real implementation, this would use the mnemonic library
        return b"example seed bytes for demonstration purposes only"
    
    def seed_to_xprv(self, seed: bytes) -> str:
        """
        Convert a seed to an extended private key (xprv)
        
        Args:
            seed: The seed as bytes
            
        Returns:
            The extended private key (xprv)
        """
        # This is a simplified implementation
        # In a real implementation, this would use a Bitcoin library
        
        # Convert seed to xprv (simplified)
        # In a real implementation, this would use a Bitcoin library
        return "xprv9s21ZrQH143K3QTDL4LXw2F7HEK3wJUD2nW2nRk4stbPy6cq3jPPqjiChkVvvNKmPGJxWUtg6LnF5kejMRNNU3TGtRBeJgk33yuGBxrMPHi" 