backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yubikey-bitcoin-seed-storage"
version = "0.1.0"
description = "YubiKey Bitcoin Seed Storage - A POC for securely storing Bitcoin seed phrases using YubiKey"
readme = "README_YUBIKEY_SALT.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "YubiKey Bitcoin Seed Storage Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "flask>=2.3.0",
    "pyyaml>=6.0.1",
    "cryptography>=41.0.7",
    "pycryptodome>=3.19.0",
    "webauthn>=1.8.0",
    "mnemonic>=0.20",
]

[tool.setuptools]
packages = [
    "api",
    "models",
    "routes",
    "services",
    "utils",
    "tests",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
